requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Pillow==10.1.0
psutil==5.9.6
zeroconf==0.131.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2==0.1.0
google-api-python-client>=2.0.0
requests==2.31.0
openai==1.3.5
SQLAlchemy-Utils==0.41.1
python-dotenv==1.0.0
paho-mqtt
qrcode
apscheduler
pytz
humanize
pyheif
sentence-transformers
demjson3
torch
ffmpeg
wand
pillow-avif-plugin
pillow-heif
pysmb>=1.2.9